chaos: fix blowfish endian problem

+++ src/chaos/cipher/blowfish.h
@@ -15,9 +15,9 @@ namespace blowfish {
 constexpr size_t kBlockSize = 8;
 constexpr size_t kMaxKeyLength = 56;
 
-void encrypt(uint32_t *key, size_t klen, uint32_t *inb, uint32_t *outb);
+void encrypt(uint8_t *key, size_t klen, uint8_t *inb, uint8_t *outb);
 
-void decrypt(uint32_t *key, size_t klen, uint32_t *inb, uint32_t *outb);
+void decrypt(uint8_t *key, size_t klen, uint8_t *inb, uint8_t *outb);
 
 }
 #endif // _BLOWFISH_H
